Add metrics

+++ src/socket/filter/config.rs
@@ -8,11 +8,11 @@ pub struct FilterConfig {
 
     /// The maximum number of requests per NodeId per second. This must be less than
     /// `max_requests_per_second`.
-    pub max_requests_per_node_per_second: Option<usize>,
+    pub max_requests_per_node_per_second: Option<f64>,
 
     /// The maximum requests tolerated per IP per second. This must be less than
     /// `max_requests_per_second`.
-    pub max_requests_per_ip_per_second: Option<usize>,
+    pub max_requests_per_ip_per_second: Option<f64>,
 }
 
 impl Default for FilterConfig {
@@ -20,8 +20,8 @@ impl Default for FilterConfig {
         FilterConfig {
             enabled: false,
             max_requests_per_second: 50,
-            max_requests_per_node_per_second: Some(10),
-            max_requests_per_ip_per_second: Some(10),
+            max_requests_per_node_per_second: Some(10.0),
+            max_requests_per_ip_per_second: Some(10.0),
         }
     }
 }
